+++ game/src/core/gui/GUIImage.java
@@ -1,8 +1,8 @@
 package core.gui;
 
 import core.gui.backend.BackendImage;
-import core.utils.math.Vector2f;
-import core.utils.math.Vector4f;
+import core.gui.math.Vector2f;
+import core.gui.math.Vector4f;
 
 public class GUIImage extends GUIElement {
