+++ game/src/core/gui/layouts/FillLayout.java
@@ -3,7 +3,7 @@ package core.gui.layouts;
 import core.gui.GUIElement;
 import core.gui.IGUILayout;
 import core.gui.layouts.hints.FillLayoutHint;
-import core.utils.math.Vector2f;
+import core.gui.math.Vector2f;
 
 import java.util.List;
