+++ game/src/core/gui/layouts/hints/AbsolutLayoutHint.java
@@ -1,7 +1,7 @@
 package core.gui.layouts.hints;
 
 import core.gui.ILayoutHint;
-import core.utils.math.Vector2i;
+import core.gui.math.Vector2i;
 
 public class AbsolutLayoutHint implements ILayoutHint {
