+++ game/src/core/gui/math/Matrix4f.java
@@ -1,4 +1,4 @@
-package core.utils.math;
+package core.gui.math;
 
 public class Matrix4f {
