+++ game/src/core/gui/math/Vector3f.java
@@ -1,4 +1,4 @@
-package core.utils.math;
+package core.gui.math;
 
 public class Vector3f {
